refactor(mouse): split the pure routing layer into mouse/route.rs (F4 Part B, 1/n)
First step of the approved decomposition in
docs/drafts/mouse-rs-decomposition-proposal.md. `src/app/mouse.rs` was
the largest file in the tree at 1,529 production lines — roughly twice
the convention and ~450 clear of the next non-guarded file.

Verbatim relocation, no behaviour change: `mouse.rs` becomes
`mouse/mod.rs`, and the pure decision layer moves to `mouse/route.rs` —
`Region`, `MouseSink`, `Gesture`, `MouseSnapshot`, `route_mouse`,
`region_at`, `hit`, `clamp_to_area`, `scroll_streak_step`,
`decide_agent_view_action` and the timing constants. `mod.rs` keeps the
ViewState data types and the impure `impl App` half, and re-exports the
pure items under their old paths so no caller changes.

route.rs is 525 lines and reads as what it is: snapshot in, decision
out, no `App`. That was the point of going first — it is the cleanest
seam, so it tests the plan before anything impure moves.

Two things worth recording for the remaining steps. The split cut
through two doc comments (the boundary landed mid-`///`), which clippy's
`empty_line_after_doc_comments` caught rather than a human — check the
seams land between items, not inside them. And `AgentViewInputs`'
fields had to become `pub` to cross the new boundary; that is the kind
of visibility creep that makes a split worse than a long file, so it is
worth watching as the impure clusters move.

mod.rs is still over 800 production lines; selection.rs, scroll.rs and
forward.rs follow.
