+++ AGENTS.md
@@ -57,7 +57,7 @@ Deep design decisions live in [`ARCHITECTURE.md`](ARCHITECTURE.md) (sync-only, M
 - Event-loop machinery — **`sources.rs`** (channel coalesce + ingest), **`loop_steps.rs`** (pre-recv drain/refresh; chord-hint settle), **`streaming.rs`** (pull-source drains; stamps `last_output_at`), **`pane_wake.rs`**, **`scheduler.rs`** (timers/deadlines incl. agent idle/anim), **`watch.rs`** (off-thread fs-watch worker).
 - Pager — **`pager_handler/`** (vi-key handling, open/close hub, modal image overlay), **`pager_stream.rs`** + **`grep_session.rs`** + **`git_view_session.rs`** (the "worker → waking channel → streaming pager" abstraction; all ride one `stream_id` / `Message::PagerStreamOutput`), **`pager_history.rs`**.
 - Pure decisions (snapshot + fn + tests) — **`route.rs`** (`route_input` → `InputSink`, shared by key + paste), **`modal.rs`** (the transient `Modal` axis), **`focus.rs`** (`decide_focus` / `recompute_focus`).
-- **`mouse.rs`** — mouse routing, the `route.rs`/`focus.rs` template applied to the pointer: a `Copy` `MouseSnapshot` + pure `route_mouse` → `MouseSink` (`Swallow`/`ListCursor`/`Pager`/`PaneForward`), plus `region_at` hit-testing the pointer against `compute_layout`'s `FrameLayout` — **the pointer, never keyboard focus**, which is what makes the wheel scroll what you're looking at. The impure half (`handle_mouse`) dispatches the sink and translates coordinates into the pane's own space before `pane::input::encode_mouse`. Buttons ride the same pure fn via a `Gesture`: **left** = focus the region (click-*through* to a mouse-aware child, reusing `set_pane_focus`/`vsplit_focus` so zoom-refusal and split-column restore come free), **middle** = `Effect::PasteFromClipboard` (routed through `handle_paste`, so bracketed-paste gating is inherited), **right** = `resolver.enter_leader()` + a due-now `chord_hint_due` so the which-key popup shows without the keyboard's debounce. Middle/right are never forwarded to a child.
+- **`mouse/`** — mouse routing. **`route.rs`** is the pure half: a `Copy` `MouseSnapshot` + pure `route_mouse` → `MouseSink`, plus `region_at` hit-testing the pointer against `compute_layout`'s `FrameLayout` — **the pointer, never keyboard focus**, which is what makes the wheel scroll what you're looking at. **`mod.rs`** holds the selection/streak data types and the impure half (`handle_mouse`), which dispatches the sink and translates coordinates into the pane's own space before `pane::input::encode_mouse`. Buttons ride the same pure fn via a `Gesture`: **left** = focus the region (click-*through* to a mouse-aware child, reusing `set_pane_focus`/`vsplit_focus` so zoom-refusal and split-column restore come free), **middle** = `Effect::PasteFromClipboard` (routed through `handle_paste`, so bracketed-paste gating is inherited), **right** = `resolver.enter_leader()` + a due-now `chord_hint_due` so the which-key popup shows without the keyboard's debounce. Middle/right are never forwarded to a child.
 - **`mouse_mode.rs`** — the terminal's mouse-reporting reconcile: `settle_mouse_mode` compares `[mouse] capture` (what the user asked for) against `view.mouse_capture_on` (what the terminal is in) at loop bottom and emits `Effect::SetMouseMode` on divergence, so startup / `:mouse` / config reload / return-from-foreground all take one path.
 - Panes — **`pane_tabs.rs`** (tab lifecycle/focus), **`pane_scroll.rs`** (scroll mode + scrollback; `^a v` auto-engages an alt-screen agent's on-disk transcript), **`codex_pin.rs`** (spawn-ordered rollout claim), **`navigate.rs`**, **`quick_select.rs`** (`^a u`).
 - Vsplit — **`vsplit.rs`** (`^a |` preview cycle; `^s n`/`^s x` second Commander; `cur()`/`col(side)`; `carve_vsplit` geometry), **`preview_ops.rs`** (off-thread live-reload of the split preview).
